This branch is now following the same form as our Application Diagram. Will add sorting and Date/Time objects in entries soon.
